+++ typescript/provider-sdk/src/chain.ts
@@ -40,6 +40,7 @@ export interface TestChainMetadata extends ChainMetadataForAltVM {
   rpcPort: number;
   rpcUrl: string;
   restPort: number;
+  gnosisSafeTransactionServiceUrl?: string;
 }
 
 /**