+++ typescript/cosmos-types/CHANGELOG.md
@@ -1,5 +1,7 @@
 # @hyperlane-xyz/cosmos-types
 
+## 19.11.0
+
 ## 19.10.0
 
 ## 19.9.0